Migrate query3d package to 64-bit floats

+++ core/spatial/query3d/aabb.go
@@ -2,16 +2,16 @@ package query3d
 
 // AABB is an axis-aligned bounding box that can be used for spatial queries.
 type AABB struct {
-	minX float32
-	minY float32
-	minZ float32
-	maxX float32
-	maxY float32
-	maxZ float32
+	minX float64
+	minY float64
+	minZ float64
+	maxX float64
+	maxY float64
+	maxZ float64
 }
 
 // NewAABB creates a new AABB with the given minimum and maximum coordinates.
-func NewAABB(minX, minY, minZ, maxX, maxY, maxZ float32) AABB {
+func NewAABB(minX, minY, minZ, maxX, maxY, maxZ float64) AABB {
 	return AABB{
 		minX: minX,
 		minY: minY,
@@ -24,7 +24,7 @@ func NewAABB(minX, minY, minZ, maxX, maxY, maxZ float32) AABB {
 
 // AABBFromSphere creates an AABB that fully contains a sphere with the given
 // center and radius.
-func AABBFromSphere(x, y, z, r float32) AABB {
+func AABBFromSphere(x, y, z, r float64) AABB {
 	return AABB{
 		minX: x - r,
 		minY: y - r,
@@ -37,7 +37,7 @@ func AABBFromSphere(x, y, z, r float32) AABB {
 
 // AABBFromBox creates an AABB that fully contains a box with the given
 // center and dimensions.
-func AABBFromBox(x, y, z, width, height, depth float32) AABB {
+func AABBFromBox(x, y, z, width, height, depth float64) AABB {
 	halfWidth := width * 0.5
 	halfHeight := height * 0.5
 	halfDepth := depth * 0.5
@@ -53,7 +53,7 @@ func AABBFromBox(x, y, z, width, height, depth float32) AABB {
 
 // AABBFromCube creates an AABB that fully contains a cube with the given
 // center and size.
-func AABBFromCube(x, y, z, size float32) AABB {
+func AABBFromCube(x, y, z, size float64) AABB {
 	halfSize := size * 0.5
 	return AABB{
 		minX: x - halfSize,
